requirements.txt
Django>=4.2,<5
gunicorn>=23,<24
PyMySQL>=1.1,<2
requests>=2.31,<3
beautifulsoup4>=4.12,<5
Pillow>=10,<12
